remade DynArr, removed stalkers feature cuz ehhh, i think if you really want to stalk on some element of array you should use list, since list is not insanely wastefull by memory(as i though, i just didnt rly had a good think about lists) only waste is some pointers but i can live with that; removed AnonDynArr becouse i dont really need it i guess; removed random .rar that got into files, sorry about that; made so everything(or almost everything) that have to do anything with pointers will use size_t as its type, i really hated size_t before for... idk it was just nonsence to me but now i get why is this a thing so i should use it since its pretty usefull and im not risking at some overflows with unsigned ints on x64 arhitectures

+++ KursovayaEngine2/Graphics/Abstractions/IndexBuffer.cpp
@@ -39,7 +39,7 @@ void IndexBufferClass::SetData(const ArrayView<void>& data) {
 		GP::IndexBufferClass::SetData(data, BufferReadWriteMode);
 	}
 }
-void IndexBufferClass::SetSubData(unsigned int offsetInBytes, const ArrayView<void>& data) {
+void IndexBufferClass::SetSubData(size_t offsetInBytes, const ArrayView<void>& data) {
 	GP::IndexBufferClass::SetSubData(offsetInBytes, data);
 }
 
@@ -51,11 +51,11 @@ void IndexBufferClass::CopyData(const IndexBufferClass& srcBuffer) {
 		GP::IndexBufferClass::CopySubData(srcBuffer, 0, 0, DataSizeInBytes);
 	}
 }
-void IndexBufferClass::CopySubData(const IndexBufferClass& srcBuffer, unsigned int srcOffsetInBytes, unsigned int dstOffsetInBytes, unsigned int amountOfBytesToCopy) {
+void IndexBufferClass::CopySubData(const IndexBufferClass& srcBuffer, size_t srcOffsetInBytes, size_t dstOffsetInBytes, size_t amountOfBytesToCopy) {
 	GP::IndexBufferClass::CopySubData(srcBuffer, srcOffsetInBytes, dstOffsetInBytes, amountOfBytesToCopy);
 }
 
-void IndexBufferClass::GetSubData(unsigned int offsetInBytes, unsigned int amountOfBytesToCopy, void* data) const {
+void IndexBufferClass::GetSubData(size_t offsetInBytes, size_t amountOfBytesToCopy, void* data) const {
 	GP::IndexBufferClass::GetSubData(offsetInBytes, amountOfBytesToCopy, data);
 }
 void IndexBufferClass::GetData(void* data) const {
